l403

+++ controllers/feed.js
@@ -1,37 +1,33 @@
 const {validationResult} = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 const path = require('path');
 const fs = require('fs');
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
-    let totalItems;
-    Post.find()
-        .countDocuments()
-        .then(count => {
-            totalItems = count;
-            return Post.find()
-                .skip((currentPage - 1) * perPage)
-                .limit(perPage);
-        })
-        .then(posts => {
-            res.status(200).json({
-                message: 'Fetched posts successfully',
-                posts: posts,
-                totalItems: totalItems
-            });
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
+    try {
+        const totalItems = await Post.find().countDocuments();
+        const posts = await Post.find()
+            .populate('creator')
+            .skip((currentPage - 1) * perPage)
+            .limit(perPage);
+        res.status(200).json({
+            message: 'Fetched posts successfully',
+            posts: posts,
+            totalItems: totalItems
         });
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
 };
 
-exports.createPost = (req, res, next) => {
+exports.createPost = async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         const error = new Error('Validation failed, entered data is incorrect.');
@@ -50,25 +46,30 @@ exports.createPost = (req, res, next) => {
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {name: 'Maximilian'}
+        creator: req.userId
     });
-    post
-        .save()
-        .then(result => {
-            res.status(201).json({
-                message: 'Post created successfully!',
-                post: result
-            });
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
+    let creator;
+    try {
+        await post.save();
+        const user = await User.findById(req.userId);
+        creator = user;
+        user.posts.push(post);
+        await user.save();
+        res.status(201).json({
+            message: 'Post created successfully!',
+            post: post,
+            creator: {
+                _id: creator._id,
+                name: creator.name
             }
-            next(err);
         });
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
 };
-//
-// imageUrl = req.file.path.replace("\\","/");
 
 exports.getPost = (req, res, next) => {
     const postId = req.params.postId;
@@ -125,6 +126,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -155,9 +161,20 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
-
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             res.status(200).json({message: 'Post deleted.'});
